+++ backend/flask_start.sh
@@ -1,4 +1,11 @@
+#!/bin/bash
+set -e
+
+# Source Redis URL construction
+. /app/construct-connection-strings.sh
+
 set -o errexit
 set -o nounset
-cd /app
-gunicorn -b :5000 app:flask_app --log-level INFO
+
+# Start Gunicorn
+exec gunicorn -b :5000 app:flask_app --log-level INFO